issue/664: success quant_linear
